[build] move bundle zip to `$(AndroidToolchainCacheDirectory)`
Previously the bundle was being downloaded to `bin/BuildDebug`, which
would be commonly cleaned for each build on CI systems. We can speed up
our builds by downloading the file to `~/android-archives`, which is
already used for various other downloads. This will persist the file
between builds on the same build agent.