[build] Properly set `$(HOME)` on Windows (#879)
Commit 6fd330df tried to provide a reasonable default for `$(HOME)`
when building on Windows (...years before we could do that):

> I've arbitrary decided to download the files into
> $(HOME)\android-archives and install them into
> $(HOME)\android-toolchain. On windows, this is
> %HOMEDRIVE%%HOMEPATH%\android-archives and
> %HOMEDRIVE%%HOMEPATH%\android-toolchain.

Where did `%HOMEDRIVE%` and `%HOMEPATH%` come from? They came from an
available Windows machine, and otherwise looked sane and useful.

Unfortunately, [when building in the VSTS build agent][0], neither
`%HOMEDRIVE%` nor `%HOMEPATH%` are set, resulting in `$(HOME)`
not being set at all!

[0]: https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=1004861

Instead of setting `$(HOME)` to `%HOMEDRIVE%%HOMEPATH%` on Windows,
set it to `%USERPROFILE`, which appears to be a more useful
environment variable.