[xa-prep-tasks] `DownloadUri` needs to create directory
Context in #880

Since we are now downloading the bundle to `~/android-archives`, it is
possible that the directory will not exist on a clean machine. We should
be calling `Directory.CreateDirectory` in the `DownloadUri` task anyway,
as this could easily happen on any destination file.

A case this occured happened on VSTS:
https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=1006149