[Xamarin.Android.Build.Tasks] Fixed Managed DesignTimeBuild.
OK completely forgot when writting this that under VS2017..
`DesignTimeBuild` is NOT set :(. As a result when the old
`Choose` statements run.. we always get a value of `False`
for `ManagedDesignTimeBuild`.

So lets rework the task to make sure we do all the required
calculations in `_SetupDesignTimeBuildForCompile` which does
get run right at the start of the `Compile`.