[Xamarin.Android.Build.Tasks] dx.jar command line too long. (#888)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=59651

We were passing ALL the `.class` files in as parameters to the
`dx.jar` invocation. Turns out we can just pass the top level
directory. In this case `$(IntermediateOutputPath)android\bin\classes`
will do fine. Because it's a directory `dx` will pick up any
`.class` file it finds in any subdirectory.

I did try this with the `.jar` files as well, but because a lot
of the libraries have a `lib` folder which contain an internal `.jar`
file it causes problems. This is because it produces duplicate types.
So rather than specifically ignore the `lib` folder (which might change)
we will just stick to the current system.