[Xamarin.Android.Build.Tests] Fix Tests using the same folders.
Many of the tests were using the same folders. This will cause
problems when we parallelize the tests as the tests will
conflict with each other. One test might delete an entire
directory causing the other to fail.

Many of the tests also take parameters, we usually use

	`TestContext.CurrentContext.Test.Name`

when we want to use the Test method as the name. However
with parameterized tests this can sometimes include invalid
characters for paths/filenames. So a new property has been
introduced in the `BaseTest` class `TestName` which cleans up
and removes problematic characters.

We also switch over to `ParallelScope.Children` so that we can
run all tests in parallel when we upgrade NUnit to 3.7.