[Xamarin.Android.Build.Tests] Use NUnit 3.7; Parallel tests (#899)
Upgrade to use NUnit 3.7, as this version supports running test
methods in parallel, not just test fixtures.

Add `[Parallelizable (ParallelScope.Children)]` where appropriate
so that tests *are* run in parallel. This cuts PR build times from
roughly 1h16min to 52min, a ~32% reduction in build times!

Fix tests so that they *can* be run in parallel. Many of the tests
were using the same folders, which causes problems when the tests are
run in parallel as the tests will conflict with each other --
one test might delete an entire directory causing another to fail.

Many of the tests also take parameters, we usually use

	TestContext.CurrentContext.Test.Name

when we want to use the Test method as the name. However
with parameterized tests this can sometimes include invalid
characters for paths/filenames. So a new property has been
introduced in the `BaseTest` class `TestName` which cleans up
and removes problematic characters.

Update the `make run-nunit-tests` target so that it now takes a
`$(TEST)` variable, which allows running a specific test fixture
or method, not all tests:

	make run-nunit-tests \
	  TEST=Xamarin.Android.Build.Tests.BuildTest.BuildBasicApplication \
	  NUNIT_TESTS=bin/TestDebug/Xamarin.Android.Build.Tests.dll