[Xamarin.Android.Build.Tests] Reworks tests to use msbuild.
This commit reworks the tests to work on msbuild by default.
xbuild is deprecated, so we should ensure that our system works
100% on msbuild.

So allot of changes in the system here. The hightlights are

We have to NOT use `BuildingInsideVisualStudio` when building
under `msbuild`. Setting this property to `true` has a side effect
of telling `msbuild` .. "Dont bother to build ProjectReferences".
This is because under VS that is something the IDE takes care of.
It turns out that the bug we had to do with `__NonExistantFile__`
causing the `csc` task to run ALL the time was fixed.

All of the `IsTargetSkipped` checks have been unified to call
`IsTargetSkipped` method out BuildOutput. It has also been updated
to include more target "skipped" formats which xbuild and msbuild
supports.

There is also a difference in behaviour between xbuild and msbuild.
You may well have noticed that ALL the tests that have had the

	"debug" -> "release"

changes (including the debugger attribute one) are around

	$(DebugType) == "None"

Under xbuild when $(DebugType) == "None" and $(DebugSymbols) == "True"
we end up with $(DebugType) == "None" and $(DebugSymbols) == "True".
Which results in a debug runtime and apk.

However.... under msbuild when $(DebugType) == "None"
and $(DebugSymbols) == "True" we end up with $(DebugType) == "portable"
and $(DebugSymbols) == "False".. so we get a release runtime and apk.

This is because None is NOT a valid value for $(DebugType) in msbuild,
so it resets BOTH $(DebugType) and $(DebugSymbols). So we have a
difference of behaviour between the two systems.

So in this commit we check for $(DebugType) == "None" and change it
and $(DebugSymbols) to match the msbuild behaviour. This will probably
break some QA tests.. but this is expected and the tests should be
updated to reflect this.