[Xamarin.Android.Bcl-Tests] Run in Release configuration
The [Jenkins `xamarin-android` build][xae9] for commit e9daf5ea was
unexpectedly UNSTABLE, because an error was encountered:

[xae9]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/662/

	Tool …/adb execution started with arguments: -s emulator-5570  install "…/xamarin-android/tests/../bin/TestRelease/Xamarin.Android.Bcl_Tests-Signed.apk"
	adb: error: cannot stat '…/xamarin-android/tests/../bin/TestRelease/Xamarin.Android.Bcl_Tests-Signed.apk':
	No such file or directory

The error was encountered because `tests/Xamarin.Android.Bcl-Tests`
wasn't built for the Release configuration, because it wasn't part of
`$(TEST_APK_PROJECTS_RELEASE)` within `Makefile`.

This raises an interesting question: what is the best solution?

 1. Make `Xamarin.Android.Bcl-Tests` a Debug-only unit test.
 2. Build and run `Xamarin.Android.Bcl-Tests` in both Debug & Release.

...quickly followed by a *second* question:

Why are some tests Debug-only and some aren't?

I don't have a good handle on this second question, other than that it
would "duplicate" unit test execution, thus lengthening build times.

However, I'm not convinced that's a good rationale, either: the
Release runtime has optimizations (`-O0 -fno-omit-frame-pointer` Debug
vs. `-O2` Release), which could theoretically break a great many
things.  The Release runtime is also what is used in our commercial
products; the Debug runtime is largely only useful for #runtime
debugging.

We thus should want as many tests as possible to run in Release, not
as few as possible

This allows us to tentatively answer the first question: we should add
`Xamarin.Android.Bcl-Tests` to `$(TEST_APK_PROJECTS_RELEASE)` so that
it's run in *both* Debug *and* Release configurations. This in turn
will ensure that the failing `adb install` command succeeds, which
*should* allow the build to PASS, as desired.