[Xamarin.Android.Build.Tasks] Create classes.zip for dx.jar. (#911)
Commit 2d23a275 reverted a change which is required for `j8.jar`.
But the change resulted in a command line which would exceed
the max command line limits on windows.

It turns out that both `dx.jar` and `j8.jar` allow for a `zip` file
to be used to pass in the `.class` files. So lets create a zip
and refresh/update it on each build. Then use that rather than
the class directory in `<CompileToDalvik/>`.

This also fixes `JarContentBuilder` to pick up the android
sdk and ndk directories from the environment variables which
the build system pass in.