[Xamarin.Android.Build.Tasks] Only Ignore Lint Checks if its supported.
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=59861

We started to ignore the `StaticFieldLeak` lint check. But this
was introduced in 26.0.2 of tools. So if we try to ignore this
check in older versions it will error.

	Invalid id or category "StaticFieldLeak".

So we need to version the checks we ignore by default. This
commit puts these in a dictionary which includes the version
number you can run them from.