[Xamarin.Android.Build.Tasks] AndroidSdk GetPlatformDirectoryFromApiLevel returns null
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=59962

`GetPlatformDirectoryFromApiLevel` is returning null. All of
the places we are expecting this to return just the path even
if it does not exist. So rather than checking if it exists in
this method we will just return the information we get.