[bcl-tests] fixes for compilation on Windows
Context:
https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=1047263

There are a couple fixes to get the BCL tests (and
Xamarin.Android-Tests.sln) compiling on Windows.

The `SystemUnzip` task is passing some file paths to
`Directory.EnumerateFiles`. Since it does not appear that this would
work; in general, it seems that `GetExtractedSourceEntries` should only
return directories as a simple fix to this problem.

Next, MSBuild is getting a duplicate `<Compile />` entry passed to `Csc`
for `obj/Debug/App.cs`. There was a conditional `<Compile />` entry as
well as one added during the `_GenerateApp_cs` task. It seems simpler to
just remove the condition and the duplicate `<Compile />` entry, since
`<BuildDependsOn />` causes the `_GenerateApp_cs` task to run first.