[tests] Add link description xml file for Bcl_tests
Some of the tests use reflection to access BCL API. Parts of the
needed API is linked away by the linker. The link description file
makes sure the needed parts stay in the mscorlib and thus the tests
don't fail.

This fixes remaining issues in the Bcl_tests in Release configuration
and the test should run without failures when linked (LinkSdkOnly).