[Xamarin.Android.Build.Tasks] Add support for `apksigner` (#928)
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=57914

This commit adds support for the `apksigner` tool which
ships with android. It is enabled by a new property

	$(AndroidUseApkSigner)

When set to true the old `jarsigner` code will be bypassed.
The new system examines the manifest file to figure out the
min and max supported versions of android. This is then passed
to the `apksigner` along with the other information (like keys etc)
and allows `apksigner` to figure out what signing algorithm to use.

Additional parameters can be passed to the tool by the developer
via the `$(AndroidApkSignerAdditionalArguments)` property.