[Xamarin.Android.Build.Tasks] Updates for xabuild.exe
This commit fixes a couple of issues which appeared when
using xabuild.exe on macos. Firstly if the symlink exists
already the SymbolicLink code will error out and fail the
build completely. This happens because if you run two or
more xabuild.exe instances at the same time they trip over
each other trying to create the symlink.

So rather than erroring completely we should only error if
the link does NOT exist. If it does exist after an attempted
creation we should ignore the exception.

The other is about where we look for extensions. Xamarin.Android
is installed on mac into

	/Library/Frameworks/Mono.framework/External/xbuild

this was not included in the search path for MSbuild, so it never
manages to find the required .targets.