[performance measurements] Fix result files location
Do not use `TestApk.ResultsPath` as base for performance measurements
anymore - it has changed recently by
https://github.com/xamarin/xamarin-android/commit/385699a58635d87a1cc65b60995d3420995f21d7. Instead
pass own results filename to the ProcessLogcatTiming and
ProcessPlotInput tasks.

Also added LabelSuffix parameter to the task, so that it is easier to
create merged measurements. It comes handy as more tests are run in
multiple configurations now. It also simplifies the use of definitions
files as we don't need one per configuration anymore.