[xa-prep-tasks] Support single-digit versions (#939)
JDK 9 reports a single-digit version value:

	$ javac -version
	javac 9

Unfortunately, this means that the `<Which/>` task won't properly
recognize JDK 9's `javac` as passing the
`%(RequiredProgram.MinimumVersion)` value of 1.8, because `<Which/>`
requires *at least* two digits, not one. (Because `System.Version`
required at least two digits; one would throw `ArgumentException`.)

Update the `<Which/>` task so that it supports single-digit version
values, treating them as if they had a "minor" value of `0`.