[build] Bump mxe and add commit sha to AndroidMxeInstallPrefix
This ensures that we can revert mxe to an earlier version.

Background:

The current "do we need to rebuild <dependency>" is mostly "forward-only".
It works properly if you bump forward in time, but if you go _backwards_,
it won't rebuild because it's timestamp based.

For example: if we bump MXE, it'll rebuild the (cached!) MXE on the build server
but if we find that MXE is broken and we need to revert... we can't because
what we're reverting to is older than the bump!

--

The new mxe now includes a few more "native" plugins in `plugins/native/<OS_SHORT_NAME>`.
This caused issues on OSX where the `patch` that mxe built ran into
a problem because it relied on newer APIs not available in older OSX.

To fix this we just disable building the native plugins and rely
on the system ones to be working. This is what mxe is doing on Travis:
https://github.com/mxe/mxe/blob/16acc77a3d45425d5a7e11d8b2511fc5a28dd56f/.travis.yml#L6