[Xamarin.Android.Build.Tests] Support MSBuild (#912)
This commit reworks the tests to pass on MSBuild. xbuild is
deprecated, so we should ensure that our system works on MSBuild.

Lots of changes. Highlights include:

We have to NOT use `$(BuildingInsideVisualStudio)` when building
under `msbuild`. Setting this property to `true` has a side effect
of telling `msbuild` "Dont bother to build @(ProjectReferences)".
This is because under VS that is something the IDE takes care of.
It turns out that the bug we had to do with `__NonExistantFile__`
causing the `csc` task to run ALL the time was fixed.

All of the `IsTargetSkipped` checks have been unified to call the
`BuildOutput.IsTargetSkipped()` method. It has also been updated
to include more target "skipped" formats which xbuild and MSBuild
generate.

There is also a difference in behaviour between xbuild and MSBuild.
You may well have noticed that ALL the tests that have had the

	"debug" -> "release"

changes (including the debugger attribute one) are around

	$(DebugType) == "None"

Under xbuild when `$(DebugType)` == "None" and
`$(DebugSymbols)` == "True", we end up with
`$(DebugType)` == "None" and `$(DebugSymbols)` == "True" (no changes).
This results in a debug runtime and apk.

However, under MSBuild when `$(DebugType)` == "None"
and `$(DebugSymbols)` == "True" we end up with
`$(DebugType)` == "portable" and `$(DebugSymbols)` == "False" (!).
This results in a release runtime and apk.

This is because `None` is NOT a valid value for `$(DebugType)` in
MSBuild, so it resets BOTH `$(DebugType)` and `$(DebugSymbols)`.
This results in difference of behaviour between the two systems.

Update `Xamarin.Android.Common.targets` to check for this, and update
`$(DebugType)` and `$(DebugSymbols)` so that xbuild behaves like
MSBuild. This improves consistency (and our own sanity!).
(This may break some QA tests, but this is now expected and the tests
should be updated to reflect this.)