[Xamarin.Android.Build.Tasks] Don't get resources in DTB
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=60080
Context: 1cd582ec

It appears that the `<GetAdditionalResourcesFromAssemblies/>` task is
causing Visual Studio 2015 to hang when loading projects.

In the interests of expediency, "revert" to d15-3 behavior and make
the `<GetAdditionalResourcesFromAssemblies/>` task invocation
conditional on `$(DesignTimeBuild)` being False, i.e. only execute
`<GetAdditionalResourcesFromAssemblies/>` when performing *normal*
builds, *not* Design-Time builds.