[Xamarin.ProjectTools] Actually use MSBuild for the tests.....
Commit 8643ded reworked the msbuild tests to use msbuild. But
there was a problem with the logic.

	string.IsNullOrEmpty (useMSBuild) || useMSBuild == "0"

well the Environment variable USE_MSBUILD will ALWAYS be empty.
Because we never set it. so the logic should be

	!string.IsNullOrEmpty (useMSBuild) && useMSBuild == "0"

so that if it is set it has to be "0" to allow xbuild to be
used.

Also one of the tests was using `RunningMSBuild` before it was
calculated. So it was always false. So just added some logging to
print what we are using to the console. This forces `RunningMSBuild`
to be calculated.

One final issue was with the intermediate .pdb files being cleaned
up by `IncrementalClean`. Turns out we had a fix for the .mdb files
in `Xamarin.Android.Windows.targets`. So the solution there was to
add the same work around for .pdb files.