[Xamarin.Android.Build.Tests] Add Build Log to the Exception data.
If the build fails we get a generic message to look at the build.log.
This is a pain because the build log is HUGE! and takes ages to find the
problem.

It seems NUnit does not include the BuildLog property from the
`FailedBuildException`. So lets override the ToString method to
include the data.