[tests] Don't use obsolete NDK API levels
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=60185

Newer versions of NDK don't include API 4 anymore but two of our tests specified
that they required that particular version to build. This caused the tests to
fail because the android-4 platform path was nowhere to be found.

Bump the minimum SDK version to 9 (which is the lowest platform included in the NDK
as of now)