[build] Support $(MonoRequiredMaximumVersion) (#960)
There are circumstances when we don't want to use the latest
and greatest version of mono. On the CI servers we will want
to keep to certain versions so that we have a stable build.
But on developer machines we will want to allow the developer
to ignore such checks.

This commit adds support for `$(MonoRequiredMaximumVersion).
By default this proeprty is ignored. But on build systems we should pass:

	/p:IgnoreMaxMonoVersion=False

to ensure that the max version is applied and we run against
a stable version of mono.