[XABuild] Work around the XABuild Config issue by creating it before the tests run.
We have a problem with multiple instances of xabuild trying to
create/overrite a config file on each run. This causes all sorts
or failures.

So let just force a call to xabuild before the tests run. We can
then skip the creation of the file as it already exists. The
environment wont change between tests so there is no need to
keep creating this file over and over again.