[Xamarin.Android.Build.Tests] Clean up DebugType = "None" test changes.
Commit 8643ded9 added support for running the tests in msbuild.
Part of that change was altering the behaviour of `xbuild to
match that of msbuild. This was regarding the way they hande
the `$(DebugType) == "None"`.

The commit altered the .targets file to reset `$(DebugType)` to
`portable` if `None` is detected. It also resets `$(DebugSymbols)`
to `false` in the same senario.

As a result the the in these two tests which do the same thing
should not longer be required.