[Xamarin.Android.Build.Tasks] Null Reference Exceoption when getting android.jar
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=60324

If a android.jar file does NOT exist for a target platform we should
raise a normal error. But we currently throw a null reference
exception.

`TryGetPlatformDirectoryFromApiLevel` can return null, this is by
design. So we need to handle that in the places where its used.
And return a reasonable error message if we do get a null back.